@@ -3,19 +3,19 @@ package com.baeldung.microprofile.providers;
 import com.baeldung.microprofile.model.Book;
 import com.baeldung.microprofile.util.BookMapper;
 
-import javax.json.Json;
-import javax.json.JsonObject;
-import javax.json.JsonWriter;
-import javax.ws.rs.Produces;
-import javax.ws.rs.WebApplicationException;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.MultivaluedMap;
-import javax.ws.rs.ext.MessageBodyWriter;
-import javax.ws.rs.ext.Provider;
-import java.io.IOException;
+import jakarta.json.Json;
+import jakarta.json.JsonObject;
+import jakarta.json.JsonWriter;
 import java.io.OutputStream;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Type;
+
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.WebApplicationException;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.MultivaluedMap;
+import jakarta.ws.rs.ext.MessageBodyWriter;
+import jakarta.ws.rs.ext.Provider;
 
 @Provider
 @Produces(MediaType.APPLICATION_JSON)
@@ -33,21 +33,11 @@ public class BookMessageBodyWriter implements MessageBodyWriter<Book> {
         return 0;
     }
 
-    /**
-     * Marsahl Book to OutputStream
-     *
-     * @param book
-     * @param type
-     * @param genericType
-     * @param annotations
-     * @param mediaType
-     * @param httpHeaders
-     * @param entityStream
-     * @throws IOException
-     * @throws WebApplicationException
-     */
     @Override
-    public void writeTo(Book book, Class<?> type, Type genericType, Annotation[] annotations, MediaType mediaType, MultivaluedMap<String, Object> httpHeaders, OutputStream entityStream) throws IOException, WebApplicationException {
+    public void writeTo(
+            Book book, Class<?> type, Type genericType, Annotation[] annotations, MediaType mediaType, MultivaluedMap<String, Object> httpHeaders,
+            OutputStream entityStream
+    ) throws WebApplicationException {
         JsonWriter jsonWriter = Json.createWriter(entityStream);
         JsonObject jsonObject = BookMapper.map(book);
         jsonWriter.writeObject(jsonObject);
