@@ -2,7 +2,7 @@ package com.baeldung.microprofile.repo;
 
 import com.baeldung.microprofile.model.Book;
 
-import javax.enterprise.context.ApplicationScoped;
+import jakarta.enterprise.context.ApplicationScoped;
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
