@@ -1,25 +1,25 @@
 package com.baeldung.hibernate.primarykeyjoincolumn;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-import org.junit.After;
-import org.junit.Before;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityManagerFactory;
 import jakarta.persistence.Persistence;
 
-public class PrimaryKeyJoinColumnIntegrationTest {
+class PrimaryKeyJoinColumnIntegrationTest {
 
     private static EntityManagerFactory emf;
 
-    private EntityManager em;
+    private static EntityManager em;
 
-    @Before
-    public void setup() {
-        emf = Persistence.createEntityManagerFactory("com.baeldung.movie_catalog");
+    @BeforeAll
+    public static void setup() {
+        emf = Persistence.createEntityManagerFactory("com.baeldung.department_person");
         em = emf.createEntityManager();
         em.getTransaction()
           .begin();
@@ -34,19 +34,18 @@ public class PrimaryKeyJoinColumnIntegrationTest {
           .commit();
     }
 
-    @After
-    public void teardown() {
+    @AfterAll
+    public static void teardown() {
         em.close();
         emf.close();
     }
 
     @Test
-    public void givenPersonEntity_getDepartment_shouldExist() {
+    void givenPersonEntity_getDepartment_shouldExist() {
         Person person = em.find(Person.class, 1L);
         assertNotNull(person);
         assertEquals("John Doe", person.getName());
         assertNotNull(person.getDepartment());
-        assertEquals("IT", person.getDepartment()
-         .getName());
+        assertEquals("IT", person.getDepartment().getName());
     }
 }
