@@ -1,5 +1,6 @@
 package com.baeldung.suanshu;
 
+/*
 import com.numericalmethod.suanshu.algebra.linear.matrix.doubles.Matrix;
 import com.numericalmethod.suanshu.algebra.linear.matrix.doubles.matrixtype.dense.DenseMatrix;
 import com.numericalmethod.suanshu.algebra.linear.matrix.doubles.operation.Inverse;
@@ -9,13 +10,14 @@ import com.numericalmethod.suanshu.analysis.function.polynomial.Polynomial;
 import com.numericalmethod.suanshu.analysis.function.polynomial.root.PolyRoot;
 import com.numericalmethod.suanshu.analysis.function.polynomial.root.PolyRootSolver;
 import com.numericalmethod.suanshu.number.complex.Complex;
+*/
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.List;
 
 class SuanShuMath {
-
+/** fixing in JAVA-24004
     private static final Logger log = LoggerFactory.getLogger(SuanShuMath.class);
 
     public static void main(String[] args) throws Exception {
@@ -137,5 +139,5 @@ class SuanShuMath {
         List<? extends Number> roots = solver.solve(p);
         log.info("Finding polynomial roots: {}", roots);
     }
-
+*/
 }
