@@ -1,15 +1,11 @@
 package com.baeldung.cassecuredapp.config;
 
 import org.jasig.cas.client.session.SingleSignOutFilter;
-import org.jasig.cas.client.validation.Cas30ServiceTicketValidator;
-import org.jasig.cas.client.validation.TicketValidator;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Primary;
 import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.ProviderManager;
 import org.springframework.security.cas.ServiceProperties;
 import org.springframework.security.cas.authentication.CasAuthenticationProvider;
 import org.springframework.security.cas.web.CasAuthenticationEntryPoint;
@@ -17,16 +13,12 @@ import org.springframework.security.cas.web.CasAuthenticationFilter;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.core.authority.AuthorityUtils;
-import org.springframework.security.core.userdetails.User;
 import org.springframework.security.web.AuthenticationEntryPoint;
+import org.springframework.security.web.SecurityFilterChain;
 import org.springframework.security.web.authentication.logout.LogoutFilter;
 
-import java.util.Collections;
-
 @EnableWebSecurity
-public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
+public class WebSecurityConfig {
 
     private Logger logger = LoggerFactory.getLogger(WebSecurityConfig.class);
     private SingleSignOutFilter singleSignOutFilter;
@@ -45,31 +37,28 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
     }
 
 
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
+    @Bean
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
         http.authorizeRequests().antMatchers( "/secured", "/login").authenticated()
-          .and()
-          .exceptionHandling().authenticationEntryPoint(authenticationEntryPoint())
-          .and()
-          .addFilterBefore(singleSignOutFilter, CasAuthenticationFilter.class)
-          .addFilterBefore(logoutFilter, LogoutFilter.class)
-          .csrf().ignoringAntMatchers("/exit/cas");
-    }
-
-    @Override
-    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-        auth.authenticationProvider(casAuthenticationProvider);
+        .and()
+        .exceptionHandling().authenticationEntryPoint(authenticationEntryPoint())
+        .and()
+        .addFilterBefore(singleSignOutFilter, CasAuthenticationFilter.class)
+        .addFilterBefore(logoutFilter, LogoutFilter.class)
+        .csrf().ignoringAntMatchers("/exit/cas");
+        return http.build();
     }
 
     @Bean
-    @Override
-    protected AuthenticationManager authenticationManager() throws Exception {
-        return new ProviderManager(Collections.singletonList(casAuthenticationProvider));
+    public AuthenticationManager authManager(HttpSecurity http) throws Exception {
+        return http.getSharedObject(AuthenticationManagerBuilder.class)
+            .authenticationProvider(casAuthenticationProvider)
+            .build();
     }
 
     public AuthenticationEntryPoint authenticationEntryPoint() {
         CasAuthenticationEntryPoint entryPoint = new CasAuthenticationEntryPoint();
-        entryPoint.setLoginUrl("https://localhost:8443/login");
+        entryPoint.setLoginUrl("https://localhost:8443/cas/login");
         entryPoint.setServiceProperties(serviceProperties);
         return entryPoint;
     }
