@@ -27,13 +27,14 @@ public class QueryService {
         return entityManager.createQuery(query).getResultList();
     }
 
-    public BookRecord findBookById(Long id) {
+    public BookRecord findBookByTitle(String title) {
         TypedQuery<BookRecord> query = entityManager
-          .createQuery("SELECT new com.baeldung.recordswithjpa.records.BookRecord(b.id, b.title, b.author, b.isbn) " +
-                  "FROM Book b WHERE b.id = :id", BookRecord.class);
-        query.setParameter("id", id);
+            .createQuery("SELECT new com.baeldung.recordswithjpa.records.BookRecord(b.id, b.title, b.author, b.isbn) " +
+                         "FROM Book b WHERE b.title = :title", BookRecord.class);
+        query.setParameter("title", title);
         return query.getSingleResult();
     }
+
 
     public List<BookRecord> findAllBooksUsingMapping() {
         Query query = entityManager.createNativeQuery("SELECT * FROM book", "BookRecordMapping");
