@@ -1,13 +1,10 @@
 package com.baeldung.keycloak;
 
-import java.security.Principal;
 import java.util.Map;
 
-import org.keycloak.KeycloakPrincipal;
-import org.keycloak.KeycloakSecurityContext;
-import org.keycloak.adapters.springsecurity.token.KeycloakAuthenticationToken;
-import org.keycloak.representations.IDToken;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.oauth2.core.oidc.OidcIdToken;
+import org.springframework.security.oauth2.core.oidc.user.DefaultOidcUser;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -18,34 +15,27 @@ public class CustomUserAttrController {
     @GetMapping(path = "/users")
     public String getUserInfo(Model model) {
 
-        KeycloakAuthenticationToken authentication = (KeycloakAuthenticationToken) SecurityContextHolder.getContext()
-            .getAuthentication();
-
-        final Principal principal = (Principal) authentication.getPrincipal();
+        final DefaultOidcUser user = (DefaultOidcUser) SecurityContextHolder.getContext()
+            .getAuthentication()
+            .getPrincipal();
 
         String dob = "";
-        String userIdByToken = "";
-        String userIdByMapper = "";
+        String userId = "";
 
-        if (principal instanceof KeycloakPrincipal) {
+        OidcIdToken token = user.getIdToken();
 
-            KeycloakPrincipal<KeycloakSecurityContext> kPrincipal = (KeycloakPrincipal<KeycloakSecurityContext>) principal;
-            IDToken token = kPrincipal.getKeycloakSecurityContext()
-                .getIdToken();
+        Map<String, Object> customClaims = token.getClaims();
 
-            userIdByToken = token.getSubject();
-            userIdByMapper = token.getOtherClaims().get("user_id").toString();
-
-            Map<String, Object> customClaims = token.getOtherClaims();
-
-            if (customClaims.containsKey("DOB")) {
-                dob = String.valueOf(customClaims.get("DOB"));
-            }
+        if (customClaims.containsKey("user_id")) {
+            userId = String.valueOf(customClaims.get("user_id"));
         }
 
-        model.addAttribute("username", principal.getName());
-        model.addAttribute("userIDByToken", userIdByToken);
-        model.addAttribute("userIDByMapper", userIdByMapper);
+        if (customClaims.containsKey("DOB")) {
+            dob = String.valueOf(customClaims.get("DOB"));
+        }
+
+        model.addAttribute("username", user.getName());
+        model.addAttribute("userID", userId);
         model.addAttribute("dob", dob);
         return "userInfo";
     }
