@@ -11,7 +11,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import com.baeldung.spring.cloud.azure.keyvault.service.KeyVaultAutoconfiguredClient;
 
 @SpringBootTest(classes = Application.class)
-public class KeyVaultAutoconfiguredClientIntegrationTest {
+class KeyVaultAutoconfiguredClientIntegrationTest {
 
     @Autowired
     @Qualifier(value = "KeyVaultAutoconfiguredClient")
@@ -22,5 +22,4 @@ public class KeyVaultAutoconfiguredClientIntegrationTest {
         String secretKey = "mySecret";
         Assertions.assertThrows(NoSuchElementException.class, () -> keyVaultAutoconfiguredClient.getSecret(secretKey));
     }
-
 }
